# resources/swagger.yaml
openapi: 3.0.0
info:
  title: Автоматизированная система управления требованиями (АСУТр)
  description: |
    API для системы управления требованиями согласно ТЗ ГОСТ 34.602-2020.
    Система предоставляет функционал для управления требованиями, их версионированием,
    согласованием, трассировкой и интеграцией со смежными системами.
  version: 1.0.0
  contact:
    name: АО "МИФИСТ"
    url: https://mifist.example.com
    email: support@mifist.example.com
servers:
  - url: https://api.asut.example.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
...
paths:
...
  /requirements:
    get:
      tags: [Требования]
      summary: Получить список требований с фильтрами
      description: Получение списка требований с возможностью фильтрации по типу, проекту и статусу
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: query
          description: Фильтр по проекту
          schema:
            type: integer
            minimum: 1
          example: 1
        - name: type
          in: query
          description: Фильтр по типу требования
          schema:
            type: string
            enum: [functional, non-functional, business]
          example: "functional"
        - name: status
          in: query
          description: Фильтр по статусу требования
          schema:
            type: string
            enum: [draft, review, approved, rejected, completed]
          example: "approved"
        - name: created_by
          in: query
          description: Фильтр по создателю требования
          schema:
            type: integer
            minimum: 1
          example: 1
        - name: priority
          in: query
          description: Фильтр по приоритету
          schema:
            type: integer
            minimum: 1
            maximum: 5
            example: 1

      responses:
        '200':
          description: Список требований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementListResponse'
        '400':
          description: Некорректные параметры фильтрации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Недостаточно прав для просмотра требований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Требования]
      summary: Создать требование
      description: Создание нового требования с валидацией данных
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementCreate'
      responses:
        '201':
          description: Требование создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementFull'
...
  /requirements/{requirement_id}/traceability:
    get:
      tags: [Требования]
      summary: Матрица трассируемости
      description: Получение матрицы трассируемости требования (другие тербваония, тест-кейсы, написать селект, который бв это описывал)
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
        - name: depth
          in: query
          schema:
            type: integer
            default: 2
      responses:
        '200':
          description: Матрица трассируемости
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceabilityMatrix'

  # ================ Покрытие тест-кейсами ================
  /requirements/{requirement_id}/test-coverage:
    get:
      tags: [Тестирование]
      summary: Получить покрытие тест-кейсами
      description: Получение информации о покрытии требования тест-кейсами
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Покрытие тест-кейсами
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCoverage'

    post:
      tags: [Тестирование]
      summary: Обновить покрытие тест-кейсами
      description: Обновление информации о покрытии тест-кейсами
      security:
        - bearerAuth: []
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCoverageUpdate'
      responses:
        '200':
          description: Покрытие обновлено

  # ================ Отчетность и аналитика ================
...
  /reports/kpi:
    get:
      tags: [Администрирование]
      summary: Дашборд KPI
      description: Получение дашборда с ключевыми показателями эффективности
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: query
          schema:
            type: integer
        - name: period_start
          in: query
          schema:
            type: string
            format: date
        - name: period_end
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Дашборд KPI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KPIDashboard'
  /reports/requirements:
    get:
      tags: [Администрирование]
      summary: Отчет по требованиям
      description: Генерация отчета по требованиям
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: query
          required: true
          schema:
            type: integer
        - name: format
          in: query
          schema:
            type: string
            enum: [json, pdf, csv]
            default: json
        - name: include_history
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Отчет по требованиям
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementsReport'
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            text/csv:
              schema:
                type: string
                format: binary
  /reports/traceability:
    get:
      tags: [Администрирование]
      summary: Отчет по трассируемости
      description: Генерация отчета по трассируемости требований
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Отчет по трассируемости
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceabilityReport'

  # ================ Администрирование ================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    # ================ Пользователи ================
...
    Requirement:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        name:
          type: string
        type:
          type: string
        created_at:
          type: string
          format: date-time
        created_by_user_id:
          type: integer
        path:
          type: string
        depth:
          type: integer
        requirement_group_id:
          type: integer
    RequirementFull:
      allOf:
        - $ref: '#/components/schemas/Requirement'
        - type: object
          properties:
            content:
              $ref: '#/components/schemas/RequirementContent'
            workflow:
              $ref: '#/components/schemas/RequirementWorkflow'
            dependencies:
              type: array
              items:
                $ref: '#/components/schemas/Dependency'
    RequirementCreate:
      type: object
      required:
        - project_id
        - name
        - description_text
      properties:
        project_id:
          type: integer
        name:
          type: string
          maxLength: 255
        type:
          type: string
        parent_id:
          type: integer
        content:
          $ref: '#/components/schemas/RequirementContentCreate'
...
    RequirementListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RequirementFull'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ================ Содержимое требований ================
    RequirementContent:
      type: object
      properties:
        id:
          type: integer
        development_basis:
          type: string
        development_purpose:
          type: string
        description_text:
          type: string
        acceptance_criteria:
          type: string
        document_requires:
          type: string
        requirement_id:
          type: integer
    RequirementContentCreate:
      type: object
      properties:
        development_basis:
          type: string
        development_purpose:
          type: string
        description_text:
          type: string
        acceptance_criteria:
          type: string
        document_requires:
          type: string
...
    RequirementWorkflow:
      type: object
      properties:
        id:
          type: integer
        requirement_content_id:
          type: integer
        priority:
          type: integer
          minimum: 1
          maximum: 5
        status:
          type: string
          enum: [draft, review, approved, rejected, implemented, verified, archived]
...
    Dependency:
      type: object
      properties:
        id:
          type: integer
        target_requirement_id:
          type: integer
        source_requirement_id:
          type: integer
        type:
          type: string
          enum: [blocks, depends_on, related_to, duplicates, refines]
        strength:
          type: integer
          minimum: 1
          maximum: 5
        description:
          type: string
        source_requirement_group_id:
          type: integer
        target_requirement_group_id:
          type: integer
...
    TraceabilityMatrix:
      type: object
      properties:
        requirement_id:
          type: integer
        requirement_name:
          type: string
        source:
          type: array
          items:
            $ref: '#/components/schemas/TraceabilityItem'
        target:
          type: array
          items:
            $ref: '#/components/schemas/TraceabilityItem'
    TraceabilityItem:
      type: object
      properties:
        requirement_id:
          type: integer
        requirement_name:
          type: string

    # ================ Тестирование ================
    TestCoverage:
      type: object
      properties:
        test_case_version_id:
          type: string
        requirement_id:
          type: integer
        created_by_user_id:
          type: integer
        created_at:
          type: string
          format: date-time
        test_case_status:
          type: string
          enum: [passed, failed, blocked, not_executed]
    TestCoverageUpdate:
      type: object
      required:
        - test_case_version_id
        - test_case_status
      properties:
        test_case_version_id:
          type: string
        test_case_status:
          type: string
          enum: [passed, failed, blocked, not_executed]
...
    KPIDashboard:
      type: object
      properties:
        period:
          type: object
          properties:
            start:
              type: string
              format: date
            end:
              type: string
              format: date
        metrics:
          type: object
          properties:
            total_requirements:
              type: integer
            requirements_approved:
              type: integer
            requirements_approval:
              type: integer
            requirements_rejected:
              type: integer
    RequirementsReport:
      type: object
      properties:
        project_id:
          type: integer
        project_name:
          type: string
        generated_at:
          type: string
          format: date-time
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/RequirementFull'
        summary:
          type: object
          properties:
            total:
              type: integer
            by_type:
              type: object
            by_status:
              type: object
    TraceabilityReport:
      type: object
      properties:
        project_id:
          type: integer
        project_name:
          type: string
        generated_at:
          type: string
          format: date-time
        traceability_matrix:
          type: array
          items:
            $ref: '#/components/schemas/TraceabilityReportItem'
    TraceabilityReportItem:
      type: object
      properties:
        requirement_id:
          type: integer
        requirement_name:
          type: string
        linked_to:
          type: array
          items:
            type: integer
        linked_from:
          type: array
          items:
            type: integer
        test_coverage:
          type: boolean

    # ================ Роли и права ================
...